merging: sort repos by id in compound shards
We sort shards in compound shards by repository ID so that we can use
binary search to quickly find a specific repo when calling
`build.findShard`.

We have seen that for very large compound shards it can take a long time
to find a specific repo which can slow down initial indexing even for
no-op indexes after a restart of indexserver.

The sort order will impact ranking of non-bm25 search, because doc-order acts
as a tie-breaker. However, the actual impact is likely very small,
because "LastCommitDate" is a stronger signal.

Test plan:
New unit test