repoID bitmap for speeding up findShard in compound shards
We add a new section to shards which contains a roaring bitmap for
quickly checking if a shard contains a repo ID. We then can load just
this (small amount) of data to rule out a compound shard. We use roaring
bitmaps since we already have that dependency in our codebase.

The reason we speed up this operation is we found on a large instance
which contained thousands of tiny repos we spent so much time in
findShard that our indexing queue would always fall behind.

It is possible this new section won't speed this up enough and we need
some sort of global oracle (or in-memory cache in indexserver?). This is
noted in the code for future travellers.

Test Plan: the existing unit tests already cover if this is forwards and
backwards compatible. Additionally I added some logging to zoekt to test
if older version of shards still work correctly in findShard, as well as
if older versions of zoekt can read the new shards.

I haven't quantified the performance improvements yet. Before landing I
will generate some synthetic large compound shards and test on my
machine the changes in perf for findShard (will update description).