ranking: remove IDF from BM25
We remove IDF because we want to use BM25 scoring for keyword search and
our query-time calculation of IDF won't work anymore if terms are
AND-ed (keyword search) instead of OR-ed (codycontext).

Our evaluations show a slight improvement which supports the decision to
treat IDF as constant.

This is also in line with how we calculate line-based BM25.

Test plan:
updated unit tests