index: decide between tenant and non-tenant shard name in one place (#953)
We have two places with duplicated logic around how it decides the layout of
shards on disk. This now moves that decision into one place.

Additionally we can now unexport index.ShardName. It was only used in one
place outside the package, and that was easy to replace with a hardcoded
string since it is just a test.

Test Plan: Just CI. This has no actual change in functionality, just
refactoring.