index: remove Options.{TenantID,RepoID} (#954)
These are duplicated with RepositoryDescription. There is exactly one
place we set these values, and in that case we set it in the repository
description as well.

Note: we update some the calls in builder.go for o.IndexOptions.RepoID
to be just o.RepoID. This is to make it consistent with how we access
TenantID. IndexOptions is embedded in that struct (indexArgs) which is
why we can do the shortcut.

Test Plan: go test is sufficient for this change. I will do a larger
round of manual tests soon with all my changes stacked.