make cache thread safe
I missed this at first so I followed up with a commit for ya. At first I
just added RWMutex in, but it was a bit complicated due to Get's needing
to do writes. So I changed the eviction policy from LRU to random. Given
you will be tuning your cache to be big enough for your use case I
imagine this will be alright?

https://ampcode.com/threads/T-10ce9d15-0e69-4903-b99e-c38459fa9219