cmd/zoekt: add support for jsonl output
Currently the zoekt command's output isn't able to surface information
like which branch a match corresponds to. By streaming out the matches
as lines of json, it can be easily processed by other programs like jq.

This is potentially useful for a lot of cli tools. Personally I'm
looking to use this with fzf and bat to build a little search tui.